Modularized Code
- Split code into 3 different files: main.py, FlightAnalysis.py, and
FlightGraphing.py
